moved selector determination to QueryHandlerFactory
