+++ src/Marten/Linq/QueryHandlers/QueryHandlerFactory.cs
@@ -4,11 +4,13 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using Baseline;
+using Marten.Events;
 using Marten.Linq.Compiled;
 using Marten.Linq.Model;
 using Marten.Linq.QueryHandlers.CompiledInclude;
 using Marten.Schema;
 using Marten.Services.Includes;
+using Marten.Transforms;
 using Npgsql;
 using Remotion.Linq;
 using Remotion.Linq.Clauses.ResultOperators;
@@ -24,6 +26,8 @@ namespace Marten.Linq.QueryHandlers
         IQueryHandler<T> BuildHandler<T>(QueryModel model, IIncludeJoin[] joins, QueryStatistics stats);
 
         IQueryHandler<TOut> HandlerFor<TDoc, TOut>(ICompiledQuery<TDoc, TOut> query);
+
+        ISelector<T> BuildSelector<T>(QueryModel query, IIncludeJoin[] joins, QueryStatistics stats);
     }
 
     public class QueryHandlerFactory : IQueryHandlerFactory
@@ -219,6 +223,71 @@ namespace Marten.Linq.QueryHandlers
             visitor.Visit(expression);
             var parameterSetters = visitor.ParameterSetters;
             return parameterSetters;
+        }
+
+        public ISelector<T> BuildSelector<T>(QueryModel query,
+    IIncludeJoin[] joins, QueryStatistics stats)
+        {
+            var mapping = _schema.MappingFor(query).ToQueryableDocument();
+            var selector = buildSelector<T>(_schema, mapping, query);
+
+            if (stats != null)
+            {
+                selector = new StatsSelector<T>(stats, selector);
+            }
+
+            if (joins.Any())
+            {
+                selector = new IncludeSelector<T>(_schema, selector, joins);
+            }
+
+            return selector;
+        }
+
+        private static ISelector<T> buildSelector<T>(IDocumentSchema schema, IQueryableDocument mapping,
+            QueryModel query)
+        {
+            var selectable = query.AllResultOperators().OfType<ISelectableOperator>().FirstOrDefault();
+            if (selectable != null)
+            {
+                return selectable.BuildSelector<T>(schema, mapping);
+            }
+
+
+            if (query.HasSelectMany())
+            {
+                return new SelectManyQuery(mapping, query, 0).ToSelector<T>(schema.StoreOptions.Serializer());
+            }
+
+            if (query.SelectClause.Selector.Type == query.SourceType())
+            {
+                if (typeof(T) == typeof(string))
+                {
+                    return (ISelector<T>)new JsonSelector();
+                }
+
+                // I'm so ashamed of this hack, but "simplest thing that works"
+                if (typeof(T) == typeof(IEvent))
+                {
+                    return mapping.As<EventQueryMapping>().Selector.As<ISelector<T>>();
+                }
+
+                if (typeof(T) != query.SourceType())
+                {
+                    // TODO -- going to have to come back to this one.
+                    return null;
+                }
+
+                var resolver = schema.ResolverFor<T>();
+
+                return new WholeDocumentSelector<T>(mapping, resolver);
+            }
+
+
+            var visitor = new SelectorParser(query);
+            visitor.Visit(query.SelectClause.Selector);
+
+            return visitor.ToSelector<T>(schema, mapping);
         }
     }
 }
